Demon Boss
Just a couple sprites for the demon Boss. It's not done and the damaged versions of the sprites aren't in yet but I figured it's good to put something in. The separate arms are supposed to attach at the bottom of the elbow. I just didn't attach it so we could turn and animate it